backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
pymongo
python-dotenv
requests
pydantic
moviepy
python-multipart
google-genai
